Port to firefox55's webextensions
Ports the extension to web extensions, so that it (mostly) runs in the
mobile firefox as well, and will continue to work with firefox 57.

There is a new settings page without much ceremony.
The settings page does not work in firefox 55 mobile for me, but works
on nightly, so I would just wait it out (no idea what's wrong).

The extension does not need to poll every few seconds to check whether
tabs can be discarded anymore.
Instead, it responds on the appropriate events (e.g. tab created, focus
changed, etc.).

Once firefox 56 is released, we can use browser.tabs.Tab.lastAccessed
instead of keeping track of it in a separate map ourselves.
I wouldn't be surprised if the current version leaked memory somewhere
because the map that keeps track of the lastAccessed property is not
cleaned property when tabs are closed.
It works fair enough for a few weeks I guess.
WeakMap`s do not seem to work with browser.tabs.Tab objects.

Relatedly, this drops support for the "do not close unseen tabs"
feature.
It could be implemented using the same mechanism used now for the
lastAccessed attribute, but I don't see how it could be implemented
without keeping a map of unseen or seen tabs of some sort.
I don't use it personally, and I don't know how important it is to
others.
I'd rather not do this and keep the extension as small and lean as
possible (there is already TabWrangler if you're into bloat).